Задание 2.4 Долматов

+++ 02-java-types/task04/src/com/example/task04/Task04.java
@@ -11,7 +11,7 @@ public class Task04 {
             case ("*"):
                 return a * b;
             case ("/"):
-                return a / b;
+                return (float)(a / b);
 
         }
         return 0;
